add model_interfaces clean up repo structure
