+++ phygnn/model_interfaces/__init__.py
@@ -1,0 +1,3 @@
+# -*- coding: utf-8 -*-
+"""Model Interfaces"""
+from .tf_model import TfModel
