+++ phygnn/model_interfaces/base_model.py
@@ -1,0 +1,852 @@
+# -*- coding: utf-8 -*-
+"""
+Base Model Interface
+"""
+import logging
+import numpy as np
+import pandas as pd
+from warnings import warn
+
+from phygnn.utilities.pre_processing import PreProcess
+
+logger = logging.getLogger(__name__)
+
+
+class ModelBase:
+    """
+    Base Model Interface
+    """
+
+    def __init__(self, model, feature_names=None, label_names=None,
+                 norm_params=None):
+        """
+        Parameters
+        ----------
+        model : OBJ
+            Sci-kit learn or tensorflow model
+        feature_names : list
+            Ordered list of feature names.
+        label_names : list
+            Ordered list of label (output) names.
+        norm_params : dict, optional
+            Dictionary mapping feature and label names (keys) to normalization
+            parameters (mean, stdev), by default None
+        """
+        self._model = model
+
+        if isinstance(feature_names, str):
+            feature_names = [feature_names]
+        elif isinstance(feature_names, np.ndarray):
+            feature_names = feature_names.tolist()
+
+        self._feature_names = feature_names
+
+        if isinstance(label_names, str):
+            label_names = [label_names]
+        elif isinstance(label_names, np.ndarray):
+            label_names = label_names.tolist()
+
+        self._label_names = label_names
+        if norm_params is None:
+            norm_params = {}
+
+        self._norm_params = norm_params
+
+    def __repr__(self):
+        msg = "{}:\n{}".format(self.__class__.__name__, self.model_summary)
+
+        return msg
+
+    def __getitem__(self, features):
+        """
+        Use model to predict label from given features
+
+        Parameters
+        ----------
+        features : pandas.DataFrame
+            features to predict from
+
+        Returns
+        -------
+        pandas.DataFrame
+            label prediction
+        """
+        return self.predict(features)
+
+    @property
+    def model_summary(self):
+        """
+        Tensorflow model summary
+
+        Returns
+        -------
+        str
+        """
+        try:
+            summary = self._model.summary()
+        except ValueError:
+            summary = None
+
+        return summary
+
+    @property
+    def feature_names(self):
+        """
+        List of the feature variable names.
+
+        Returns
+        -------
+        list
+        """
+        return self._feature_names
+
+    @property
+    def feature_dims(self):
+        """
+        Number of features
+
+        Returns
+        -------
+        int
+        """
+        n_features = (len(self.feature_names)
+                      if self.feature_names is not None else None)
+
+        return n_features
+
+    @property
+    def label_names(self):
+        """
+        label variable names
+
+        Returns
+        -------
+        list
+        """
+        return self._label_names
+
+    @property
+    def label_dims(self):
+        """
+        Number of labels
+
+        Returns
+        -------
+        int
+        """
+        n_labels = (len(self.label_names)
+                    if self.label_names is not None else None)
+
+        return n_labels
+
+    @property
+    def normalization_parameters(self):
+        """
+        Features and label (un)normalization parameters
+
+        Returns
+        -------
+        dict
+        """
+        return self._norm_params
+
+    @property
+    def means(self):
+        """
+        Mapping feature/label names to the mean values for
+        (un)normalization
+
+        Returns
+        -------
+        dict
+        """
+        means = {k: v['mean'] for k, v in self._norm_params.items()}
+
+        return means
+
+    @property
+    def stdevs(self):
+        """
+        Mapping feature/label names to the stdev values for
+        (un)normalization
+
+        Returns
+        -------
+        dict
+        """
+        stdevs = {k: v['stdev'] for k, v in self._norm_params.items()}
+
+        return stdevs
+
+    @property
+    def model(self):
+        """
+        Trained model
+
+        Returns
+        -------
+        tensorflow.keras.models
+        """
+        return self._model
+
+    @property
+    def feature_means(self):
+        """
+        Feature means, used for (un)normalization
+
+        Returns
+        -------
+        list
+        """
+        means = None
+        if self._feature_names is not None:
+            means = []
+            for f in self._feature_names:
+                means.append(self.get_mean(f))
+
+        return means
+
+    @property
+    def feature_stdevs(self):
+        """
+        Feature stdevs, used for (un)normalization
+
+        Returns
+        -------
+        list
+        """
+        stdevs = None
+        if self._feature_names is not None:
+            stdevs = []
+            for f in self._feature_names:
+                stdevs.append(self.get_stdev(f))
+
+        return stdevs
+
+    @property
+    def label_means(self):
+        """
+        label means, used for (un)normalization
+
+        Returns
+        -------
+        list
+        """
+        means = None
+        if self.label_names is not None:
+            means = []
+            for l_n in self.label_names:
+                means.append(self.get_mean(l_n))
+
+        return means
+
+    @property
+    def label_stdevs(self):
+        """
+        label stdevs, used for (un)normalization
+
+        Returns
+        -------
+        list
+        """
+        stdevs = None
+        if self.label_names is not None:
+            stdevs = []
+            for l_n in self.label_names:
+                stdevs.append(self.get_stdev(l_n))
+
+        return stdevs
+
+    @staticmethod
+    def _normalize(native_arr, mean=None, stdev=None):
+        """
+        Normalize features with mean at 0 and stdev of 1.
+
+        Parameters
+        ----------
+        native_arr : ndarray
+            native data
+        mean : float | None
+            mean to use for normalization
+        stdev : float | None
+            stdev to use for normalization
+
+        Returns
+        -------
+        norm_arr : ndarray
+            normalized data
+        mean : float
+            mean used for normalization
+        stdev : float
+            stdev used for normalization
+        """
+
+        if mean is None:
+            mean = np.nanmean(native_arr, axis=0)
+
+        if stdev is None:
+            stdev = np.nanstd(native_arr, axis=0)
+
+        norm_arr = native_arr - mean
+        norm_arr /= stdev
+
+        return norm_arr, mean, stdev
+
+    @staticmethod
+    def _unnormalize(norm_arr, mean, stdev):
+        """
+        Unnormalize data with mean at 0 and stdev of 1.
+
+        Parameters
+        ----------
+        norm_arr : ndarray
+            normalized data
+        mean : float
+            mean used for normalization
+        stdev : float
+            stdev used for normalization
+
+        Returns
+        -------
+        native_arr : ndarray
+            native un-normalized data
+        """
+        native_arr = norm_arr * stdev
+        native_arr += mean
+
+        return native_arr
+
+    @staticmethod
+    def dict_json_convert(inp):
+        """Recursively convert numeric values in dict to work with json dump
+
+        Parameters
+        ----------
+        inp : dict
+            Dictionary to convert.
+
+        Returns
+        -------
+        out : dict
+            Copy of dict input with all nested numeric values converted to
+            base python int or float and all arrays converted to lists.
+        """
+
+        if isinstance(inp, dict):
+            out = {k: ModelBase.dict_json_convert(v) for k, v in inp.items()}
+        elif isinstance(inp, (list, tuple)):
+            out = [ModelBase.dict_json_convert(i) for i in inp]
+        elif np.issubdtype(type(inp), np.floating):
+            out = float(inp)
+        elif np.issubdtype(type(inp), np.integer):
+            out = int(inp)
+        elif isinstance(inp, np.ndarray):
+            out = inp.tolist()
+        else:
+            out = inp
+
+        return out
+
+    @staticmethod
+    def _parse_data(data, names=None):
+        """
+        Parse data array and names from input data
+
+        Parameters
+        ----------
+        data : pandas.DataFrame | dict | ndarray
+            Features/labels to parse
+        names : list, optional
+            List of data item names, by default None
+
+        Returns
+        -------
+        data : ndarray
+            Data array
+        names: list
+            List of data item names
+        """
+        if isinstance(data, pd.DataFrame):
+            names = data.columns.values.tolist()
+            data = data.values
+        elif isinstance(data, dict):
+            names = list(data.keys())
+            data = np.dstack(list(data.values()))[0]
+        elif isinstance(data, np.ndarray):
+            if names is None:
+                msg = ('Names of items must be supplied to parse data '
+                       'arrays')
+                logger.error(msg)
+                raise RuntimeError(msg)
+
+        return data, names
+
+    def get_norm_params(self, names):
+        """
+        Get means and stdevs for given feature/label names
+
+        Parameters
+        ----------
+        names : list
+            list of feature/label names to get normalization params for
+
+        Returns
+        -------
+        means : list
+            List of means to use for (un)normalization
+        stdevs : list
+            List of stdevs to use for (un)normalization
+        """
+        means = []
+        stdevs = []
+        for name in names:
+            means.append(self.get_mean(name))
+            stdevs.append(self.get_stdev(name))
+
+        if not all(set(means)):
+            means = None
+
+        if not all(set(stdevs)):
+            stdevs = None
+
+        return means, stdevs
+
+    def get_mean(self, name):
+        """
+        Get feature | label mean
+
+        Parameters
+        ----------
+        name : str
+            feature | label name
+
+        Returns
+        -------
+        mean : float
+            Mean value used for normalization
+        """
+        mean = self._norm_params.get(name, None)
+        if mean is not None:
+            mean = mean.get('mean', None)
+
+        return mean
+
+    def get_stdev(self, name):
+        """
+        Get feature | label stdev
+
+        Parameters
+        ----------
+        name : str
+            feature | label name
+
+        Returns
+        -------
+        stdev : float
+            Stdev value used for normalization
+        """
+        stdev = self._norm_params.get(name, None)
+        if stdev is not None:
+            stdev = stdev.get('stdev', None)
+
+        return stdev
+
+    def _normalize_dict(self, items):
+        """
+        Normalize given dictionary of items (features | labels)
+
+        Parameters
+        ----------
+        items : dict
+            mapping of names to vectors
+
+        Returns
+        -------
+        norm_items : dict
+            mapping of names to normalized-feature vectors
+        """
+        norm_items = {}
+        for key, value in items.items():
+            mean = self.get_mean(key)
+            stdev = self.get_stdev(key)
+            update = mean is None or stdev is None
+            try:
+                value, mean, stdev = self._normalize(value, mean=mean,
+                                                     stdev=stdev)
+                if update:
+                    norm_params = {key: {'mean': mean, 'stdev': stdev}}
+                    self._norm_params.update(norm_params)
+            except Exception as ex:
+                msg = "Could not normalize {}:\n{}".format(key, ex)
+                logger.warning(msg)
+                warn(msg)
+
+            norm_items[key] = value
+
+        return norm_items
+
+    def _normalize_df(self, df):
+        """
+        Normalize DataFrame
+
+        Parameters
+        ----------
+        df : pandas.DataFrame
+            DataFrame of features/label to normalize
+
+        Returns
+        -------
+        norm_df : pandas.DataFrame
+            Normalized features/label
+        """
+        means, stdevs = self.get_norm_params(df.columns)
+        update = means is None or stdevs is None
+
+        norm_df, means, stdevs = self._normalize(df, mean=means,
+                                                 stdev=stdevs)
+        if update:
+            for i, c in enumerate(df.columns):
+                norm_params = {c: {'mean': means[i], 'stdev': stdevs[i]}}
+                self._norm_params.update(norm_params)
+
+        return norm_df
+
+    def _normalize_arr(self, arr, names):
+        """
+        Normalize array and save normalization parameters to given names
+
+        Parameters
+        ----------
+        arr : ndarray
+            Array of features/label to normalize
+        names : list
+            List of feature/label names
+
+        Returns
+        -------
+        norm_arr : ndarray
+            Normalized features/label
+        """
+        if len(names) != arr.shape[1]:
+            msg = ("Number of item names ({}) does not match number of items "
+                   "({})".format(len(names), arr.shape[1]))
+            logger.error(msg)
+            raise RuntimeError(msg)
+
+        means, stdevs = self.get_norm_params(names)
+        update = means is None or stdevs is None
+
+        norm_arr, means, stdevs = self._normalize(arr, mean=means,
+                                                  stdev=stdevs)
+        if update:
+            for i, n in enumerate(names):
+                norm_params = {n: {'mean': means[i], 'stdev': stdevs[i]}}
+                self._norm_params.update(norm_params)
+
+        return norm_arr
+
+    def normalize(self, data, names=None):
+        """
+        Normalize given data
+
+        Parameters
+        ----------
+        data : dict | pandas.DataFrame | ndarray
+            Data to normalize
+        names : list, optional
+            List of data item names, needed to normalized ndarrays,
+            by default None
+
+        Returns
+        -------
+        data : dict | pandas.DataFrame | ndarray
+            Normalized data in same format as input
+        """
+        if isinstance(data, dict):
+            data = self._normalize_dict(data)
+        elif isinstance(data, pd.DataFrame):
+            data = self._normalize_df(data)
+        elif isinstance(data, (list, np.ndarray)):
+            if names is None:
+                msg = ('Names of items must be supplied to nomralize data '
+                       'arrays')
+                logger.error(msg)
+                raise RuntimeError(msg)
+            else:
+                data = self._normalize_arr(data, names)
+        else:
+            msg = "Cannot normalize data of type: {}".format(type(data))
+            logger.error(msg)
+            raise RuntimeError(msg)
+
+        return data
+
+    def _unnormalize_dict(self, items):
+        """
+        Un-normalize given dictionary of items (features | labels)
+
+        Parameters
+        ----------
+        items : dict
+            mapping of names to vectors
+
+        Returns
+        -------
+        native_items : dict
+            mapping of names to native vectors
+        """
+        native_items = {}
+        for key, value in items.items():
+            norm_params = self.normalization_parameters[key]
+            if norm_params is not None:
+                value = self._unnormalize(value, norm_params['mean'],
+                                          norm_params['stdev'])
+            else:
+                msg = ("Normalization Parameters unavailable for {}"
+                       .format(key))
+                logger.warning(msg)
+                warn(msg)
+
+            native_items[key] = value
+
+        return native_items
+
+    def _unnormalize_df(self, df):
+        """
+        Un-normalize DataFrame
+
+        Parameters
+        ----------
+        df : pandas.DataFrame
+            DataFrame of features/label to un-normalize
+
+        Returns
+        -------
+        native_df : pandas.DataFrame
+            Native features/label array
+        """
+        means, stdevs = self.get_norm_params(df.columns)
+
+        native_df = self._unnormalize(df, means, stdevs)
+
+        return native_df
+
+    def _unnormalize_arr(self, arr, names):
+        """
+        Un-normalize array using given names
+
+        Parameters
+        ----------
+        arr : ndarray
+            Array of features/label to un-normalize
+        names : list
+            List of feature/label names
+
+        Returns
+        -------
+        native_arr : ndarray
+            Native features/label array
+        """
+        if len(names) != arr.shape[1]:
+            msg = ("Number of item names ({}) does not match number of items "
+                   "({})".format(len(names), arr.shape[1]))
+            logger.error(msg)
+            raise RuntimeError(msg)
+
+        means, stdevs = self.get_norm_params(names)
+
+        native_arr = self._unnormalize(arr, means, stdevs)
+
+        return native_arr
+
+    def unnormalize(self, data, names=None):
+        """
+        Un-normalize given data
+
+        Parameters
+        ----------
+        data : dict | pandas.DataFrame | ndarray
+            Data to un-normalize
+        names : list, optional
+            List of data item names, needed to un-normalized ndarrays,
+            by default None
+
+        Returns
+        -------
+        data : dict | pandas.DataFrame | ndarray
+            Native data in same format as input
+        """
+        if isinstance(data, dict):
+            data = self._unnormalize_dict(data)
+        elif isinstance(data, pd.DataFrame):
+            data = self._unnormalize_df(data)
+        elif isinstance(data, (list, np.ndarray)):
+            if names is None:
+                msg = ('Names of items must be supplied to un-nomralize data '
+                       'arrays')
+                logger.error(msg)
+                raise RuntimeError(msg)
+            else:
+                data = self._unnormalize_arr(data, names)
+        else:
+            msg = "Cannot un-normalize data of type: {}".format(type(data))
+            logger.error(msg)
+            raise RuntimeError(msg)
+
+        return data
+
+    def _check_one_hot_norm_params(self, one_hot_features):
+        """
+        Check one hot feature normalization parameters to ensure they are
+        {mean: 0, stdev: 1} to prevent normalization
+
+        Parameters
+        ----------
+        one_hot_features : list
+            list of one hot features
+        """
+        for feature in one_hot_features:
+            mean = self.get_mean(feature)
+            stdev = self.get_stdev(feature)
+            if mean != 0 and stdev != 1:
+                norm_params = {feature: {'mean': 0, 'stdev': 1}}
+                self._norm_params.update(norm_params)
+
+    def _parse_features(self, features, names=None, process_one_hot=True,
+                        **kwargs):
+        """
+        Parse features
+
+        Parameters
+        ----------
+        features : pandas.DataFrame | dict | ndarray
+            Features to train on or predict from
+        names : list, optional
+            List of feature names, by default None
+        process_one_hot : bool, optional
+            Check for and process one-hot variables, by default True
+        kwargs : dict, optional
+            kwargs for PreProcess.one_hot
+
+        Returns
+        -------
+        features : ndarray
+            Parsed features array normalized and with str columns converted
+            to one hot vectors if desired
+        """
+        if len(features.shape) != 2:
+            msg = ('{} can only use 2D data as input!'
+                   .format(self.__class__.__name__))
+            logger.error(msg)
+            raise RuntimeError(msg)
+
+        if self.feature_names is not None:
+            if features.shape[1] != len(self.feature_names):
+                msg = ('data has {} features but expected {}'
+                       .format(features.shape[1], self.feature_dims))
+                logger.error(msg)
+                raise RuntimeError(msg)
+
+        features, feature_names = self._parse_data(features, names=names)
+
+        if self._feature_names is None:
+            self._feature_names = feature_names
+        elif self.feature_names != feature_names:
+            msg = ('Expecting features with names: {}, but was provided with: '
+                   '{}!'.format(feature_names, self.feature_names))
+            logger.error(msg)
+            raise RuntimeError(msg)
+
+        if process_one_hot:
+            kwargs.update({'return_ind': True})
+            features, one_hot_ind = PreProcess.one_hot(features, **kwargs)
+            self._check_one_hot_norm_params(self.feature_names[one_hot_ind])
+
+        features = self.normalize(features, names=feature_names)
+
+        return features
+
+    def _parse_labels(self, labels, names=None, normalize=True):
+        """
+        Parse labels and normalize if desired
+
+        Parameters
+        ----------
+        labels : pandas.DataFrame | dict | ndarray
+            Features to train on or predict from
+        names : list, optional
+            List of label names, by default None
+        normalize : bool, optional
+            Normalize label array, by default True
+
+        Returns
+        -------
+        labels : ndarray
+            Parsed labels array, normalized if desired
+        """
+        if self.label_names is not None:
+            if len(labels.shape) == 1:
+                n_labels = len(labels)
+            else:
+                n_labels = labels.shape[1]
+
+            if n_labels != len(self.label_names):
+                msg = ('data has {} labels but expected {}'
+                       .format(labels.shape[1], self.label_dims))
+                logger.error(msg)
+                raise RuntimeError(msg)
+
+        labels, label_names = self._parse_data(labels, names=names)
+
+        if self._label_names is None:
+            self._label_names = label_names
+        elif self.label_names != label_names:
+            msg = ('Expecting labels with names: {}, but was provided with: '
+                   '{}!'.format(label_names, self.label_names))
+            logger.error(msg)
+            raise RuntimeError(msg)
+
+        if normalize:
+            labels = self.normalize(labels, names=label_names)
+
+        return labels
+
+    def predict(self, features, table=True, parse_kwargs=None,
+                predict_kwargs=None):
+        """
+        Use model to predict label from given features
+
+        Parameters
+        ----------
+        features : dict | pandas.DataFrame
+            features to predict from
+        table : bool, optional
+            Return pandas DataFrame
+        parse_kwargs : dict
+            kwargs for cls._parse_features
+        predict_wargs : dict
+            kwargs for tensorflow.*.predict
+
+        Returns
+        -------
+        prediction : ndarray | pandas.DataFrame
+            label prediction
+        """
+        if parse_kwargs is None:
+            parse_kwargs = {}
+
+        if isinstance(features, np.ndarray):
+            parse_kwargs.update({"names": self.feature_names})
+
+        features = self._parse_features(features, **parse_kwargs)
+
+        if predict_kwargs is None:
+            predict_kwargs = {}
+
+        prediction = self._model.predict(features, **predict_kwargs)
+        prediction = self.unnormalize(prediction, names=self.label_names)
+
+        if table:
+            prediction = pd.DataFrame(prediction, columns=self.label_names)
+
+        return prediction
