+++ phygnn/utilities/tf_utilities.py
@@ -1,3 +1,4 @@
+# -*- coding: utf-8 -*-
 """
 Tensorflow utilities
 """
